🧹 NEE2TR task: keep context specs under hotspot limit
